+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Summary
 * Enhancement - Improvements for the UI in the pattern screen: [#3580](https://github.com/owncloud/android/issues/3580)
 * Enhancement - Release Notes: [#3442](https://github.com/owncloud/android/issues/3442)
 * Enhancement - What´s new option: [#3352](https://github.com/owncloud/android/issues/3352)
+* Enhancement - New option to show or not hidden files: [#2578](https://github.com/owncloud/android/issues/2578)
 
 Details
 -------
@@ -91,6 +92,13 @@ Details
 
    https://github.com/owncloud/android/issues/3352
    https://github.com/owncloud/android/pull/3616
+
+* Enhancement - New option to show or not hidden files: [#2578](https://github.com/owncloud/android/issues/2578)
+
+   Enable it to show hidden files and folders
+
+   https://github.com/owncloud/android/issues/2578
+   https://github.com/owncloud/android/pull/3624
 
 Changelog for ownCloud Android Client [2.20.0] (2022-02-16)
 =======================================